@@ -29,106 +38,188 @@ def generateGitBuild = { ->
     } catch (ignored) {
         stringBuilder.append('NoGitSystemAvailable')
     }
-    stringBuilder.append('-')
-    stringBuilder.append((new Date()).format('yyyy.MM.dd'))
-    stringBuilder.append('"')
     return stringBuilder.toString()
 }
 
+def generateGitRemote = { ->
+    StringBuilder stringBuilder = new StringBuilder()
+    try {
+        def stdout = new ByteArrayOutputStream()
+        exec {
+            commandLine 'git', 'remote', 'get-url', 'origin'
+            standardOutput = stdout
+        }
+        String commitObject = stdout.toString().trim()
+        stringBuilder.append(commitObject)
+    } catch (ignored) {
+        stringBuilder.append('NoGitSystemAvailable')
+    }
+    return stringBuilder.toString()
+}
+
+def generateDate = { ->
+    StringBuilder stringBuilder = new StringBuilder()
+    stringBuilder.append((new Date()).format('yyyy.MM.dd-HH:mm'))
+    return stringBuilder.toString()
+}
+
+def isMaster = { ->
+    return !version.contains('-')
+}
+
+def gitAvailable = { ->
+    StringBuilder stringBuilder = new StringBuilder()
+    try {
+        def stdout = new ByteArrayOutputStream()
+        exec {
+            commandLine 'git', '--version'
+            standardOutput = stdout
+        }
+        String commitObject = stdout.toString().trim()
+        stringBuilder.append(commitObject)
+    } catch (ignored) {
+        return false // NoGitSystemAvailable
+    }
+    return !stringBuilder.toString().isEmpty()
+
+}
+
+def allCommitted = { ->
+    StringBuilder stringBuilder = new StringBuilder()
+    try {
+        def stdout = new ByteArrayOutputStream()
+        exec {
+            commandLine 'git', 'status', '-s'
+            standardOutput = stdout
+        }
+        // ignore all changes done in .idea/codeStyles
+        String cleanedList = stdout.toString().replaceAll(/(?m)^\s*(M|A|D|\?\?)\s*.*?\.idea\/codeStyles\/.*?\s*$/, "")
+        // ignore all files added to project dir but not staged/known to GIT
+        cleanedList = cleanedList.replaceAll(/(?m)^\s*(\?\?)\s*.*?\s*$/, "")
+        stringBuilder.append(cleanedList.trim())
+    } catch (ignored) {
+        return false // NoGitSystemAvailable
+    }
+    return stringBuilder.toString().isEmpty()
+}
+
+tasks.matching { it instanceof Test }.all {
+    testLogging.events = ["failed", "skipped", "started"]
+    testLogging.exceptionFormat = "full"
+}
+
 android {
-    compileSdkVersion 23
-    buildToolsVersion "25.0.2"
+    compileSdkVersion 28
+    ndkVersion "21.1.6352462"
 
     defaultConfig {
-        applicationId "info.nightscout.androidaps"
-        minSdkVersion 21
-        targetSdkVersion 23
+        minSdkVersion 26
+        targetSdkVersion 28
+        multiDexEnabled true
         versionCode 1500
-        version "1.5d"
+        version "2.8.2.1-dev-b"
         buildConfigField "String", "VERSION", '"' + version + '"'
-        buildConfigField "String", "BUILDVERSION", generateGitBuild()
+        buildConfigField "String", "BUILDVERSION", '"' + generateGitBuild() + '-' + generateDate() + '"'
+        buildConfigField "String", "REMOTE", '"' + generateGitRemote() + '"'
+        buildConfigField "String", "HEAD", '"' + generateGitBuild() + '"'
+        buildConfigField "String", "COMMITTED", '"' + allCommitted() + '"'
+        testInstrumentationRunner "androidx.test.runner.AndroidJUnitRunner"
+        // if you change minSdkVersion to less than 11, you need to change executeTask for wear
+
+        ndk {
+            moduleName "BleCommandUtil"
+        }
+    }
+    kotlinOptions {
+        jvmTarget = '1.8'
+    }
+    buildFeatures {
+        viewBinding true
     }
     lintOptions {
+        checkReleaseBuilds false
         disable 'MissingTranslation'
+        disable 'ExtraTranslation'
     }
     buildTypes {
         release {
             minifyEnabled false
             proguardFiles getDefaultProguardFile('proguard-android.txt'), 'proguard-rules.pro'
         }
-    }
-    productFlavors {
-        flavorDimensions "standard", "wear"
-        full {
-            dimension "standard"
-            resValue "string", "app_name", "AndroidAPS"
-            versionName version
-            manifestPlaceholders = [
-                    appIcon: "@mipmap/blueowl"
-            ]
-            buildConfigField "boolean", "APS", "true"
-            buildConfigField "boolean", "PUMPDRIVERS", "true"
-            buildConfigField "boolean", "NSCLIENTOLNY", "false"
-            buildConfigField "boolean", "CLOSEDLOOP", "true"
+        debug {
+            testCoverageEnabled(project.hasProperty('coverage'))
         }
-        openloop {
+        firebaseDisable {
+            System.setProperty("disableFirebase", "true")
+            ext.enableCrashlytics = false
+        }
+    }
+    flavorDimensions "standard"
+    productFlavors {
+        full {
+            applicationId "info.nightscout.androidaps"
             dimension "standard"
             resValue "string", "app_name", "AndroidAPS"
             versionName version
             manifestPlaceholders = [
-                    appIcon: "@mipmap/blueowl"
+                    appIcon     : "@mipmap/ic_launcher",
+                    appIconRound: "@mipmap/ic_launcher_round"
             ]
-            buildConfigField "boolean", "APS", "true"
-            buildConfigField "boolean", "PUMPDRIVERS", "true"
-            buildConfigField "boolean", "NSCLIENTOLNY", "false"
-            buildConfigField "boolean", "CLOSEDLOOP", "false"
         }
         pumpcontrol {
+            applicationId "info.nightscout.aapspumpcontrol"
             dimension "standard"
-            resValue "string", "app_name", "AndroidAPS"
-            versionName version
+            resValue "string", "app_name", "Pumpcontrol"
+            versionName version + "-pumpcontrol"
             manifestPlaceholders = [
-                    appIcon: "@mipmap/blueowl"
+                    appIcon     : "@mipmap/ic_pumpcontrol",
+                    appIconRound: "@null"
             ]
-            buildConfigField "boolean", "APS", "false"
-            buildConfigField "boolean", "PUMPDRIVERS", "true"
-            buildConfigField "boolean", "NSCLIENTOLNY", "false"
-            buildConfigField "boolean", "CLOSEDLOOP", "false"
         }
         nsclient {
+            applicationId "info.nightscout.nsclient"
             dimension "standard"
             resValue "string", "app_name", "NSClient"
             versionName version + "-nsclient"
             manifestPlaceholders = [
-                    appIcon: "@mipmap/yellowowl"
+                    appIcon     : "@mipmap/ic_yellowowl",
+                    appIconRound: "@null"
             ]
-            buildConfigField "boolean", "APS", "false"
-            buildConfigField "boolean", "PUMPDRIVERS", "false"
-            buildConfigField "boolean", "NSCLIENTOLNY", "true"
-            buildConfigField "boolean", "CLOSEDLOOP", "false"
         }
-        wear {
-            dimension "wear"
-            buildConfigField "boolean", "WEAR", "true"
-            buildConfigField "boolean", "WEAR_CONTROL", "false"
-
-        }
-        wearcontrol {
-            dimension "wear"
-            buildConfigField "boolean", "WEAR", "true"
-            buildConfigField "boolean", "WEAR_CONTROL", "true"
-        }
-        nowear {
-            dimension "wear"
-            buildConfigField "boolean", "WEAR", "false"
-            buildConfigField "boolean", "WEAR_CONTROL", "false"
+        nsclient2 {
+            applicationId "info.nightscout.nsclient2"
+            dimension "standard"
+            resValue "string", "app_name", "NSClient2"
+            versionName version + "-nsclient"
+            manifestPlaceholders = [
+                    appIcon     : "@mipmap/ic_yellowowl",
+                    appIconRound: "@null"
+            ]
         }
     }
+    compileOptions {
+        sourceCompatibility JavaVersion.VERSION_1_8
+        targetCompatibility JavaVersion.VERSION_1_8
+    }
+
+    testOptions {
+        unitTests {
+            returnDefaultValues = true
+            includeAndroidResources = true
+
+            all {
+                maxParallelForks = 10
+                forkEvery = 20
+            }
+        }
+    }
+
+    useLibrary "org.apache.http.legacy"
+
 }
 
 allprojects {
     repositories {
-        jcenter()
         flatDir {
             dirs 'libs'
         }
