@@ -1,58 +1,107 @@
-apply plugin: 'com.android.application'
+buildscript {
+    repositories {
+        google()
+        mavenCentral()
+        maven { url "https://plugins.gradle.org/m2/" } // jacoco 0.2
+    }
 
+    dependencies {
+        //classpath 'com.dicedmelon.gradle:jacoco-android:0.1.4'
+        classpath 'com.hiya:jacoco-android:0.2'
+    }
+}
+apply plugin: 'com.android.application'
+//apply plugin: 'jacoco-android'
+apply plugin: 'com.hiya.jacoco-android'
+
+jacoco {
+    toolVersion = "0.8.3"
+}
+
+ext {
+    wearableVersion = "2.8.1"
+    playServicesWearable = "17.0.0"
+}
+
+def generateGitBuild = { ->
+
+    StringBuilder stringBuilder = new StringBuilder()
+    stringBuilder.append('"')
+    try {
+        def stdout = new ByteArrayOutputStream()
+        exec {
+            commandLine 'git', 'describe', '--always'
+            standardOutput = stdout
+        }
+        String commitObject = stdout.toString().trim()
+        stringBuilder.append(commitObject)
+    } catch (ignored) {
+        stringBuilder.append('NoGitSystemAvailable')
+    }
+    stringBuilder.append('-')
+    stringBuilder.append((new Date()).format('yyyy.MM.dd'))
+    stringBuilder.append('"')
+    return stringBuilder.toString()
+}
 
 android {
-    compileSdkVersion 23
-    buildToolsVersion "25.0.2"
+    compileSdkVersion 28
 
     defaultConfig {
         applicationId "info.nightscout.androidaps"
-        minSdkVersion 20
-        targetSdkVersion 23
-        versionCode 1
-        versionName "1.0.2"
+        minSdkVersion 23
+        targetSdkVersion 28
+        versionCode 2
+        versionName "1.0.3"
+        buildConfigField "String", "BUILDVERSION", generateGitBuild()
     }
     buildTypes {
         release {
             minifyEnabled false
             proguardFiles getDefaultProguardFile('proguard-android.txt'), 'proguard-rules.pro'
         }
+        firebaseDisable {
+        }
     }
 
-    publishNonDefault true
+    compileOptions {
+        sourceCompatibility JavaVersion.VERSION_1_8
+        targetCompatibility JavaVersion.VERSION_1_8
+    }
 
+    lintOptions {
+        abortOnError false
+    }
+
+    flavorDimensions "standard"
     productFlavors {
         full {
-            applicationId = "info.nightscout.androidaps"
-            resValue "string", "label_xdrip", "AAPS"
-            resValue "string", "label_xdrip_large", "AAPS(Large)"
-            resValue "string", "label_xdrip_big_chart", "AAPS(BigChart)"
-            resValue "string", "label_xdrip_no_chart", "AAPS(NoChart)"
-            resValue "string", "label_xdrip_circle", "AAPS(Circle)"
-            resValue "string", "label_xdrip_activity", "AAPS Prefs."
-            resValue "string", "app_settings", "AAPS Settings"
-            buildConfigField "boolean", "WEAR_CONTROL", "true"
-
+            applicationId "info.nightscout.androidaps"
+            dimension "standard"
+            versionName version
         }
-
-        restricted {
-            applicationId = "info.nightscout.androidaps"
-            resValue "string", "label_xdrip", "AAPS"
-            resValue "string", "label_xdrip_large", "AAPS(Large)"
-            resValue "string", "label_xdrip_big_chart", "AAPS(BigChart)"
-            resValue "string", "label_xdrip_no_chart", "AAPS(NoChart)"
-            resValue "string", "label_xdrip_circle", "AAPS(Circle)"
-            resValue "string", "label_xdrip_activity", "AAPS Prefs."
-            resValue "string", "app_settings", "AAPS Settings"
-            buildConfigField "boolean", "WEAR_CONTROL", "false"
-
+        pumpcontrol {
+            applicationId "info.nightscout.aapspumpcontrol"
+            dimension "standard"
+            versionName version + "-pumpcontrol"
+        }
+        nsclient {
+            applicationId "info.nightscout.nsclient"
+            dimension "standard"
+            versionName version + "-nsclient"
+        }
+        nsclient2 {
+            applicationId "info.nightscout.nsclient2"
+            dimension "standard"
+            versionName version + "-nsclient"
         }
     }
 }
 
 allprojects {
     repositories {
-        jcenter()
+        google()
+        mavenCentral()
         flatDir {
             dirs 'libs'
         }
