@@ -15,8 +15,9 @@ import android.graphics.RectF;
 import android.graphics.Shader;
 import android.os.Bundle;
 import android.os.PowerManager;
+import android.os.SystemClock;
 import android.preference.PreferenceManager;
-import android.support.v4.content.LocalBroadcastManager;
+import androidx.localbroadcastmanager.content.LocalBroadcastManager;
 import android.support.wearable.watchface.WatchFaceStyle;
 import android.util.Log;
 import android.util.TypedValue;
@@ -27,17 +28,16 @@ import android.view.WindowManager;
 import android.widget.TextView;
 
 import com.google.android.gms.wearable.DataMap;
-import com.ustwo.clockwise.wearable.WatchFace;
 import com.ustwo.clockwise.common.WatchFaceTime;
+import com.ustwo.clockwise.wearable.WatchFace;
 
 import java.util.ArrayList;
 import java.util.Calendar;
-import java.util.Date;
 import java.util.HashSet;
 import java.util.TreeSet;
 
-import info.nightscout.androidaps.data.BgWatchData;
 import info.nightscout.androidaps.R;
+import info.nightscout.androidaps.data.BgWatchData;
 import info.nightscout.androidaps.interaction.menus.MainMenuActivity;
 
 
@@ -54,8 +54,8 @@ public class CircleWatchface extends WatchFace implements SharedPreferences.OnSh
     private float angleSMALL = 0f;
     private int hour, minute;
     private int color;
-    private Paint circlePaint = new Paint();
-    private Paint removePaint = new Paint();
+    private final Paint circlePaint = new Paint();
+    private final Paint removePaint = new Paint();
     private RectF rect, rectDelete;
     private boolean overlapping;
 
@@ -64,7 +64,7 @@ public class CircleWatchface extends WatchFace implements SharedPreferences.OnSh
 
 
     public Point displaySize = new Point();
-    private MessageReceiver messageReceiver = new MessageReceiver();
+    private final MessageReceiver messageReceiver = new MessageReceiver();
 
     private int sgvLevel = 0;
     private String sgvString = "999";
@@ -72,13 +72,12 @@ public class CircleWatchface extends WatchFace implements SharedPreferences.OnSh
 
 
     private int batteryLevel = 0;
-    private double datetime = 0;
+    private long datetime = 0;
     private String direction = "";
     private String delta = "";
     private String avgDelta = "";
-    public TreeSet<BgWatchData> bgDataList = new TreeSet<BgWatchData>();
+    public TreeSet<BgWatchData> bgDataList = new TreeSet<>();
 
-    private View layoutView;
     private int specW;
     private int specH;
     private View myLayout;
@@ -93,7 +92,7 @@ public class CircleWatchface extends WatchFace implements SharedPreferences.OnSh
         super.onCreate();
 
         PowerManager powerManager = (PowerManager) getSystemService(POWER_SERVICE);
-        PowerManager.WakeLock wakeLock = powerManager.newWakeLock(PowerManager.PARTIAL_WAKE_LOCK, "CreateWakelock");
+        PowerManager.WakeLock wakeLock = powerManager.newWakeLock(PowerManager.PARTIAL_WAKE_LOCK, "AndroidAPS:CircleWatchface");
         wakeLock.acquire(30000);
 
         Display display = ((WindowManager) getSystemService(Context.WINDOW_SERVICE))
@@ -150,9 +149,9 @@ public class CircleWatchface extends WatchFace implements SharedPreferences.OnSh
 
         // prepare fields
 
-        TextView textView = null;
-        mSgv = (TextView) myLayout.findViewById(R.id.sgvString);
-        textView = (TextView) myLayout.findViewById(R.id.sgvString);
+        TextView textView;
+        mSgv = myLayout.findViewById(R.id.sgvString);
+        textView = myLayout.findViewById(R.id.sgvString);
         if (sharedPrefs.getBoolean("showBG", true)) {
             textView.setVisibility(View.VISIBLE);
             textView.setText(getSgvString());
@@ -163,7 +162,7 @@ public class CircleWatchface extends WatchFace implements SharedPreferences.OnSh
             textView.setVisibility(View.INVISIBLE);
         }
 
-        textView = (TextView) myLayout.findViewById(R.id.statusString);
+        textView = myLayout.findViewById(R.id.statusString);
         if (sharedPrefs.getBoolean("showExternalStatus", true)) {
             textView.setVisibility(View.VISIBLE);
             textView.setText(getStatusString());
@@ -174,7 +173,7 @@ public class CircleWatchface extends WatchFace implements SharedPreferences.OnSh
             textView.setVisibility(View.GONE);
         }
 
-        textView = (TextView) myLayout.findViewById(R.id.agoString);
+        textView = myLayout.findViewById(R.id.agoString);
         if (sharedPrefs.getBoolean("showAgo", true)) {
             textView.setVisibility(View.VISIBLE);
 
@@ -190,7 +189,7 @@ public class CircleWatchface extends WatchFace implements SharedPreferences.OnSh
             textView.setVisibility(View.INVISIBLE);
         }
 
-        textView = (TextView) myLayout.findViewById(R.id.deltaString);
+        textView = myLayout.findViewById(R.id.deltaString);
         if (sharedPrefs.getBoolean("showDelta", true)) {
             textView.setVisibility(View.VISIBLE);
             textView.setText(getDelta());
@@ -200,7 +199,7 @@ public class CircleWatchface extends WatchFace implements SharedPreferences.OnSh
             } else {
                 textView.setTextSize(TypedValue.COMPLEX_UNIT_SP, 18);
             }
-            if(sharedPrefs.getBoolean("showAvgDelta", true)){
+            if (sharedPrefs.getBoolean("showAvgDelta", true)) {
                 textView.append("  " + getAvgDelta());
             }
 
@@ -215,9 +214,9 @@ public class CircleWatchface extends WatchFace implements SharedPreferences.OnSh
     }
 
     public String getMinutes() {
-        String minutes = "--\'";
+        String minutes = "--'";
         if (getDatetime() != 0) {
-            minutes = ((int) Math.floor((System.currentTimeMillis() - getDatetime()) / 60000)) + "\'";
+            minutes = ((int) Math.floor((System.currentTimeMillis() - getDatetime()) / 60000.0)) + "'";
         }
         return minutes;
     }
@@ -297,10 +296,8 @@ public class CircleWatchface extends WatchFace implements SharedPreferences.OnSh
         removePaint.setAntiAlias(true);
         removePaint.setColor(getBackgroundColor());
 
-        ;
-
-        rect = new RectF(PADDING, PADDING, (float) (displaySize.x - PADDING), (float) (displaySize.y - PADDING));
-        rectDelete = new RectF(PADDING - CIRCLE_WIDTH / 2, PADDING - CIRCLE_WIDTH / 2, (float) (displaySize.x - PADDING + CIRCLE_WIDTH / 2), (float) (displaySize.y - PADDING + CIRCLE_WIDTH / 2));
+        rect = new RectF(PADDING, PADDING, displaySize.x - PADDING, displaySize.y - PADDING);
+        rectDelete = new RectF(PADDING - CIRCLE_WIDTH / 2, PADDING - CIRCLE_WIDTH / 2, displaySize.x - PADDING + CIRCLE_WIDTH / 2, displaySize.y - PADDING + CIRCLE_WIDTH / 2);
         overlapping = ALWAYS_HIGHLIGT_SMALL || areOverlapping(angleSMALL, angleSMALL + SMALL_HAND_WIDTH + NEAR, angleBig, angleBig + BIG_HAND_WIDTH + NEAR);
         Log.d("CircleWatchface", "end prepareDrawTime");
 
@@ -325,11 +322,11 @@ public class CircleWatchface extends WatchFace implements SharedPreferences.OnSh
     protected void onTimeChanged(WatchFaceTime oldTime, WatchFaceTime newTime) {
         if (oldTime.hasMinuteChanged(newTime)) {
             PowerManager powerManager = (PowerManager) getSystemService(POWER_SERVICE);
-            PowerManager.WakeLock wakeLock = powerManager.newWakeLock(PowerManager.PARTIAL_WAKE_LOCK, "TimeChangedWakelock");
+            PowerManager.WakeLock wakeLock = powerManager.newWakeLock(PowerManager.PARTIAL_WAKE_LOCK, "AndroidAPS:CircleWatchface_onTimeChanged");
             wakeLock.acquire(30000);
             /*Preparing the layout just on every minute tick:
-            *  - hopefully better battery life
-            *  - drawback: might update the minutes since last reading up endTime one minute late*/
+             *  - hopefully better battery life
+             *  - drawback: might update the minutes since last reading up endTime one minute late*/
             prepareLayout();
             prepareDrawTime();
             invalidate();  //redraw the time
@@ -432,11 +429,11 @@ public class CircleWatchface extends WatchFace implements SharedPreferences.OnSh
     }
 
 
-    private synchronized double getDatetime() {
+    private synchronized long getDatetime() {
         return datetime;
     }
 
-    private synchronized void setDatetime(double datetime) {
+    private synchronized void setDatetime(long datetime) {
         this.datetime = datetime;
     }
 
@@ -475,6 +472,7 @@ public class CircleWatchface extends WatchFace implements SharedPreferences.OnSh
     private String getAvgDelta() {
         return avgDelta;
     }
+
     private void setAvgDelta(String avgDelta) {
         this.avgDelta = avgDelta;
     }
@@ -504,11 +502,7 @@ public class CircleWatchface extends WatchFace implements SharedPreferences.OnSh
                 setIsAnimated(true);
                 for (int i = 0; i <= 8 * 1000 / 40; i++) {
                     animationStep();
-                    try {
-                        Thread.sleep(40);
-                    } catch (InterruptedException e) {
-                        e.printStackTrace();
-                    }
+                    SystemClock.sleep(40);
                 }
                 setIsAnimated(false);
                 prepareDrawTime();
@@ -525,11 +519,10 @@ public class CircleWatchface extends WatchFace implements SharedPreferences.OnSh
         @Override
         public void onReceive(Context context, Intent intent) {
             PowerManager powerManager = (PowerManager) getSystemService(POWER_SERVICE);
-            PowerManager.WakeLock wakeLock = powerManager.newWakeLock(PowerManager.PARTIAL_WAKE_LOCK,
-                    "MyWakelockTag");
+            PowerManager.WakeLock wakeLock = powerManager.newWakeLock(PowerManager.PARTIAL_WAKE_LOCK, "AndroidAPS:MessageReceiver");
             wakeLock.acquire(30000);
             Bundle bundle = intent.getBundleExtra("data");
-            if (bundle!= null) {
+            if (bundle != null) {
                 DataMap dataMap = DataMap.fromBundle(bundle);
                 setSgvLevel((int) dataMap.getLong("sgvLevel"));
                 Log.d("CircleWatchface", "sgv level : " + getSgvLevel());
@@ -537,7 +530,7 @@ public class CircleWatchface extends WatchFace implements SharedPreferences.OnSh
                 Log.d("CircleWatchface", "sgv string : " + getSgvString());
                 setDelta(dataMap.getString("delta"));
                 setAvgDelta(dataMap.getString("avgDelta"));
-                setDatetime(dataMap.getDouble("timestamp"));
+                setDatetime(dataMap.getLong("timestamp"));
                 addToWatchSet(dataMap);
 
 
@@ -568,7 +561,7 @@ public class CircleWatchface extends WatchFace implements SharedPreferences.OnSh
 
     public synchronized void addToWatchSet(DataMap dataMap) {
 
-        if(!sharedPrefs.getBoolean("showRingHistory", false)){
+        if (!sharedPrefs.getBoolean("showRingHistory", false)) {
             bgDataList.clear();
             return;
         }
@@ -579,8 +572,9 @@ public class CircleWatchface extends WatchFace implements SharedPreferences.OnSh
             double sgv = dataMap.getDouble("sgvDouble");
             double high = dataMap.getDouble("high");
             double low = dataMap.getDouble("low");
-            double timestamp = dataMap.getDouble("timestamp");
-            bgDataList.add(new BgWatchData(sgv, high, low, timestamp));
+            long timestamp = dataMap.getLong("timestamp");
+            int color = dataMap.getInt("color", 0);
+            bgDataList.add(new BgWatchData(sgv, high, low, timestamp, color));
         } else if (!sharedPrefs.getBoolean("animation", false)) {
             // don't load history at once if animations are set (less resource consumption)
             Log.d("addToWatchSet", "entries.size(): " + entries.size());
@@ -589,13 +583,14 @@ public class CircleWatchface extends WatchFace implements SharedPreferences.OnSh
                 double sgv = entry.getDouble("sgvDouble");
                 double high = entry.getDouble("high");
                 double low = entry.getDouble("low");
-                double timestamp = entry.getDouble("timestamp");
-                bgDataList.add(new BgWatchData(sgv, high, low, timestamp));
+                long timestamp = entry.getLong("timestamp");
+                int color = entry.getInt("color", 0);
+                bgDataList.add(new BgWatchData(sgv, high, low, timestamp, color));
             }
         } else
 
             Log.d("addToWatchSet", "start removing bgDataList.size(): " + bgDataList.size());
-        HashSet removeSet = new HashSet();
+        HashSet<BgWatchData> removeSet = new HashSet<>();
         double threshold = (System.currentTimeMillis() - (1000 * 60 * 5 * holdInMemory()));
         for (BgWatchData data : bgDataList) {
             if (data.timestamp < threshold) {
@@ -674,7 +669,7 @@ public class CircleWatchface extends WatchFace implements SharedPreferences.OnSh
         }
 
         float offsetMultiplier = (((displaySize.x / 2f) - PADDING) / 12f);
-        float offset = (float) Math.max(1, Math.ceil((System.currentTimeMillis() - entry.timestamp) / (1000 * 60 * 5)));
+        float offset = (float) Math.max(1, Math.ceil((System.currentTimeMillis() - entry.timestamp) / (1000 * 60 * 5.0)));
         size = bgToAngle((float) entry.sgv);
         addArch(canvas, offset * offsetMultiplier + 10, color, (float) size);
         addArch(canvas, (float) size, offset * offsetMultiplier + 10, getBackgroundColor(), (float) (360 - size));
@@ -700,7 +695,7 @@ public class CircleWatchface extends WatchFace implements SharedPreferences.OnSh
             barColor = darken(getLowColor(), .5);
         }
         float offsetMultiplier = (((displaySize.x / 2f) - PADDING) / 12f);
-        float offset = (float) Math.max(1, Math.ceil((System.currentTimeMillis() - entry.timestamp) / (1000 * 60 * 5)));
+        float offset = (float) Math.max(1, Math.ceil((System.currentTimeMillis() - entry.timestamp) / (1000 * 60 * 5.0)));
         size = bgToAngle((float) entry.sgv);
         addArch(canvas, offset * offsetMultiplier + 11, barColor, (float) size - 2); // Dark Color Bar
         addArch(canvas, (float) size - 2, offset * offsetMultiplier + 11, indicatorColor, 2f); // Indicator at end of bar
@@ -711,14 +706,14 @@ public class CircleWatchface extends WatchFace implements SharedPreferences.OnSh
     @Override
     protected void onTapCommand(int tapType, int x, int y, long eventTime) {
 
-        int extra = mSgv!=null?(mSgv.getRight() - mSgv.getLeft())/2:0;
+        int extra = mSgv != null ? (mSgv.getRight() - mSgv.getLeft()) / 2 : 0;
 
-        if (tapType == TAP_TYPE_TAP&&
-                x + extra >=mSgv.getLeft() &&
-                x - extra <= mSgv.getRight()&&
+        if (tapType == TAP_TYPE_TAP &&
+                x + extra >= mSgv.getLeft() &&
+                x - extra <= mSgv.getRight() &&
                 y >= mSgv.getTop() &&
-                y <= mSgv.getBottom()){
-            if (eventTime - sgvTapTime < 800){
+                y <= mSgv.getBottom()) {
+            if (eventTime - sgvTapTime < 800) {
                 Intent intent = new Intent(this, MainMenuActivity.class);
                 intent.addFlags(Intent.FLAG_ACTIVITY_NEW_TASK);
                 startActivity(intent);
@@ -728,7 +723,7 @@ public class CircleWatchface extends WatchFace implements SharedPreferences.OnSh
     }
 
     @Override
-    protected WatchFaceStyle getWatchFaceStyle(){
+    protected WatchFaceStyle getWatchFaceStyle() {
         return new WatchFaceStyle.Builder(this).setAcceptsTapEvents(true).build();
     }
 
