@@ -1,7 +1,6 @@
 package info.nightscout.androidaps.interaction.actions;
 
 
-import android.app.Activity;
 import android.content.res.Resources;
 import android.os.Bundle;
 import android.support.wearable.view.DotsPageIndicator;
@@ -11,7 +10,6 @@ import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
 import android.widget.ImageView;
-import android.widget.TextView;
 
 import java.text.DecimalFormat;
 
@@ -35,10 +33,10 @@ public class BolusActivity extends ViewSelectorActivity {
         super.onCreate(savedInstanceState);
         setContentView(R.layout.grid_layout);
         final Resources res = getResources();
-        final GridViewPager pager = (GridViewPager) findViewById(R.id.pager);
+        final GridViewPager pager = findViewById(R.id.pager);
 
         pager.setAdapter(new MyGridViewPagerAdapter());
-        DotsPageIndicator dotsPageIndicator = (DotsPageIndicator) findViewById(R.id.page_indicator);
+        DotsPageIndicator dotsPageIndicator = findViewById(R.id.page_indicator);
         dotsPageIndicator.setPager(pager);
     }
 
@@ -64,30 +62,30 @@ public class BolusActivity extends ViewSelectorActivity {
         @Override
         public Object instantiateItem(ViewGroup container, int row, int col) {
 
-            if(col == 0){
+            if (col == 0) {
                 final View view = getInflatedPlusMinusView(container);
                 double def = 0;
-                if (editInsulin != null){
+                if (editInsulin != null) {
                     def = SafeParse.stringToDouble(editInsulin.editText.getText().toString());
                 }
                 editInsulin = new PlusMinusEditText(view, R.id.amountfield, R.id.plusbutton, R.id.minusbutton, def, 0d, 30d, 0.1d, new DecimalFormat("#0.0"), false);
-                setLabelToPlusMinusView(view, "insulin");
+                setLabelToPlusMinusView(view, getString(R.string.action_insulin));
                 container.addView(view);
                 return view;
-            } else if(col == 1){
+            } else if (col == 1) {
                 final View view = getInflatedPlusMinusView(container);
                 double def = 0;
-                if (editCarbs != null){
+                if (editCarbs != null) {
                     def = SafeParse.stringToDouble(editCarbs.editText.getText().toString());
                 }
                 editCarbs = new PlusMinusEditText(view, R.id.amountfield, R.id.plusbutton, R.id.minusbutton, def, 0d, 150d, 1d, new DecimalFormat("0"), false);
-                setLabelToPlusMinusView(view, "carbs");
+                setLabelToPlusMinusView(view, getString(R.string.action_carbs));
                 container.addView(view);
                 return view;
             } else {
 
                 final View view = LayoutInflater.from(getApplicationContext()).inflate(R.layout.action_send_item, container, false);
-                final ImageView confirmbutton = (ImageView) view.findViewById(R.id.confirmbutton);
+                final ImageView confirmbutton = view.findViewById(R.id.confirmbutton);
                 confirmbutton.setOnClickListener(new View.OnClickListener() {
                     @Override
                     public void onClick(View v) {
@@ -95,7 +93,7 @@ public class BolusActivity extends ViewSelectorActivity {
                         //check if it can happen that the fagment is never created that hold data?
                         // (you have to swipe past them anyways - but still)
 
-                        String actionstring = "bolus " +SafeParse.stringToDouble(editInsulin.editText.getText().toString())
+                        String actionstring = "bolus " + SafeParse.stringToDouble(editInsulin.editText.getText().toString())
                                 + " " + SafeParse.stringToInt(editCarbs.editText.getText().toString());
                         ListenerService.initiateAction(BolusActivity.this, actionstring);
                         finish();
@@ -110,12 +108,12 @@ public class BolusActivity extends ViewSelectorActivity {
         public void destroyItem(ViewGroup container, int row, int col, Object view) {
             // Handle this to get the data before the view is destroyed?
             // Object should still be kept by this, just setup for reinit?
-            container.removeView((View)view);
+            container.removeView((View) view);
         }
 
         @Override
         public boolean isViewFromObject(View view, Object object) {
-            return view==object;
+            return view == object;
         }
 
 
