@@ -1,9 +1,10 @@
 package info.nightscout.androidaps.interaction.actions;
 
 
-import android.app.Activity;
+import android.content.SharedPreferences;
 import android.content.res.Resources;
 import android.os.Bundle;
+import android.preference.PreferenceManager;
 import android.support.wearable.view.DotsPageIndicator;
 import android.support.wearable.view.GridPagerAdapter;
 import android.support.wearable.view.GridViewPager;
@@ -11,12 +12,11 @@ import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
 import android.widget.ImageView;
-import android.widget.TextView;
 
 import java.text.DecimalFormat;
 
-import info.nightscout.androidaps.data.ListenerService;
 import info.nightscout.androidaps.R;
+import info.nightscout.androidaps.data.ListenerService;
 import info.nightscout.androidaps.interaction.utils.PlusMinusEditText;
 import info.nightscout.androidaps.interaction.utils.SafeParse;
 
@@ -28,21 +28,23 @@ import info.nightscout.androidaps.interaction.utils.SafeParse;
 public class WizardActivity extends ViewSelectorActivity {
 
     PlusMinusEditText editCarbs;
+    PlusMinusEditText editPercentage;
 
-    boolean useBG;
-    boolean includeBolusIOB;
-    boolean includeBasalIOB;
+    boolean hasPercentage;
+
 
     @Override
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
         setContentView(R.layout.grid_layout);
         final Resources res = getResources();
-        final GridViewPager pager = (GridViewPager) findViewById(R.id.pager);
+        final GridViewPager pager = findViewById(R.id.pager);
 
         pager.setAdapter(new MyGridViewPagerAdapter());
-        DotsPageIndicator dotsPageIndicator = (DotsPageIndicator) findViewById(R.id.page_indicator);
+        DotsPageIndicator dotsPageIndicator = findViewById(R.id.page_indicator);
         dotsPageIndicator.setPager(pager);
+        SharedPreferences sp = PreferenceManager.getDefaultSharedPreferences(this);
+        hasPercentage = sp.getBoolean("wizardpercentage", false);
     }
 
 
@@ -56,7 +58,7 @@ public class WizardActivity extends ViewSelectorActivity {
     private class MyGridViewPagerAdapter extends GridPagerAdapter {
         @Override
         public int getColumnCount(int arg0) {
-            return 5;
+            return hasPercentage ? 3 : 2;
         }
 
         @Override
@@ -67,7 +69,7 @@ public class WizardActivity extends ViewSelectorActivity {
         @Override
         public Object instantiateItem(ViewGroup container, int row, int col) {
 
-            if(col == 0){
+            if (col == 0) {
                 final View view = getInflatedPlusMinusView(container);
                 if (editCarbs == null) {
                     editCarbs = new PlusMinusEditText(view, R.id.amountfield, R.id.plusbutton, R.id.minusbutton, 0d, 0d, 150d, 1d, new DecimalFormat("0"), false);
@@ -76,85 +78,24 @@ public class WizardActivity extends ViewSelectorActivity {
                     editCarbs = new PlusMinusEditText(view, R.id.amountfield, R.id.plusbutton, R.id.minusbutton, def, 0d, 150d, 1d, new DecimalFormat("0"), false);
 
                 }
-                setLabelToPlusMinusView(view, "carbs");
+                setLabelToPlusMinusView(view, getString(R.string.action_carbs));
                 container.addView(view);
                 return view;
-            } else if(col == 1){
-                final View view = LayoutInflater.from(getApplicationContext()).inflate(R.layout.action_toggle_item, container, false);
-                final TextView textView = (TextView) view.findViewById(R.id.label);
-                textView.setText("include BG?");
-
-                final ImageView togglebutton = (ImageView) view.findViewById(R.id.togglebutton);
-                if(useBG){
-                    togglebutton.setImageResource(R.drawable.ic_toggle_on);
+            } else if (col == 1 && hasPercentage) {
+                final View view = getInflatedPlusMinusView(container);
+                if (editPercentage == null) {
+                    editPercentage = new PlusMinusEditText(view, R.id.amountfield, R.id.plusbutton, R.id.minusbutton, 100d, 50d, 150d, 1d, new DecimalFormat("0"), false);
                 } else {
-                    togglebutton.setImageResource(R.drawable.ic_toggle_off);
+                    double def = SafeParse.stringToDouble(editPercentage.editText.getText().toString());
+                    editPercentage = new PlusMinusEditText(view, R.id.amountfield, R.id.plusbutton, R.id.minusbutton, def, 50d, 150d, 1d, new DecimalFormat("0"), false);
                 }
-                togglebutton.setOnClickListener(new View.OnClickListener() {
-                    @Override
-                    public void onClick(View v) {
-                        useBG = !useBG;
-                        if(useBG){
-                            togglebutton.setImageResource(R.drawable.ic_toggle_on);
-                        } else {
-                            togglebutton.setImageResource(R.drawable.ic_toggle_off);
-                        }
-                    }
-                });
-                container.addView(view);
-                return view;
-            } else if(col == 2){
-                final View view = LayoutInflater.from(getApplicationContext()).inflate(R.layout.action_toggle_item, container, false);
-                final TextView textView = (TextView) view.findViewById(R.id.label);
-                textView.setText("Bolus IOB?");
-
-                final ImageView togglebutton = (ImageView) view.findViewById(R.id.togglebutton);
-                if(includeBolusIOB){
-                    togglebutton.setImageResource(R.drawable.ic_toggle_on);
-                } else {
-                    togglebutton.setImageResource(R.drawable.ic_toggle_off);
-                }
-                togglebutton.setOnClickListener(new View.OnClickListener() {
-                    @Override
-                    public void onClick(View v) {
-                        includeBolusIOB = !includeBolusIOB;
-                        if(includeBolusIOB){
-                            togglebutton.setImageResource(R.drawable.ic_toggle_on);
-                        } else {
-                            togglebutton.setImageResource(R.drawable.ic_toggle_off);
-                        }
-                    }
-                });
-                container.addView(view);
-                return view;
-            } else if(col == 3){
-                final View view = LayoutInflater.from(getApplicationContext()).inflate(R.layout.action_toggle_item, container, false);
-                final TextView textView = (TextView) view.findViewById(R.id.label);
-                textView.setText("Basal IOB?");
-
-                final ImageView togglebutton = (ImageView) view.findViewById(R.id.togglebutton);
-                if(includeBasalIOB){
-                    togglebutton.setImageResource(R.drawable.ic_toggle_on);
-                } else {
-                    togglebutton.setImageResource(R.drawable.ic_toggle_off);
-                }
-                togglebutton.setOnClickListener(new View.OnClickListener() {
-                    @Override
-                    public void onClick(View v) {
-                        includeBasalIOB = !includeBasalIOB;
-                        if(includeBasalIOB){
-                            togglebutton.setImageResource(R.drawable.ic_toggle_on);
-                        } else {
-                            togglebutton.setImageResource(R.drawable.ic_toggle_off);
-                        }
-                    }
-                });
+                setLabelToPlusMinusView(view, getString(R.string.action_percentage));
                 container.addView(view);
                 return view;
             } else {
 
                 final View view = LayoutInflater.from(getApplicationContext()).inflate(R.layout.action_send_item, container, false);
-                final ImageView confirmbutton = (ImageView) view.findViewById(R.id.confirmbutton);
+                final ImageView confirmbutton = view.findViewById(R.id.confirmbutton);
                 confirmbutton.setOnClickListener(new View.OnClickListener() {
                     @Override
                     public void onClick(View v) {
@@ -162,10 +103,13 @@ public class WizardActivity extends ViewSelectorActivity {
                         //check if it can happen that the fagment is never created that hold data?
                         // (you have to swipe past them anyways - but still)
 
-                        String actionstring = "wizard " + SafeParse.stringToInt(editCarbs.editText.getText().toString())
-                                + " " + useBG
-                                + " " + includeBolusIOB
-                                + " " + includeBasalIOB;
+                        int percentage = 100;
+
+                        if (editPercentage != null)
+                            percentage = SafeParse.stringToInt(editPercentage.editText.getText().toString());
+
+                        String actionstring = "wizard2 " + SafeParse.stringToInt(editCarbs.editText.getText().toString())
+                                + " " + percentage;
                         ListenerService.initiateAction(WizardActivity.this, actionstring);
                         finish();
                     }
@@ -179,12 +123,12 @@ public class WizardActivity extends ViewSelectorActivity {
         public void destroyItem(ViewGroup container, int row, int col, Object view) {
             // Handle this to get the data before the view is destroyed?
             // Object should still be kept by this, just setup for reinit?
-            container.removeView((View)view);
+            container.removeView((View) view);
         }
 
         @Override
         public boolean isViewFromObject(View view, Object object) {
-            return view==object;
+            return view == object;
         }
 
 
