@@ -1,5 +1,6 @@
 package info.nightscout.androidaps.interaction.menus;
 
+import info.nightscout.androidaps.R;
 import info.nightscout.androidaps.data.ListenerService;
 import info.nightscout.androidaps.interaction.utils.MenuListActivity;
 
@@ -12,34 +13,24 @@ public class StatusMenuActivity extends MenuListActivity {
     @Override
     protected String[] getElements() {
         return new String[] {
-                "Pump",
-                "Loop",
-                "Targets",
-                "CPP",
-                "OAPS Result"};
+                getString(R.string.status_pump),
+                getString(R.string.status_loop),
+                getString(R.string.status_cpp),
+                getString(R.string.status_tdd)};
+
 
     }
 
     @Override
-    protected void doAction(int position) {
-        switch (position) {
-
-            case 0:
-                ListenerService.initiateAction(this, "status pump");
-                break;
-            case 1:
-                ListenerService.initiateAction(this, "status loop");
-                break;
-            case 2:
-                ListenerService.initiateAction(this, "status targets");
-                break;
-            case 3:
-                ListenerService.initiateAction(this, "opencpp");
-                break;
-            case 4:
-                ListenerService.initiateAction(this, "status oapsresult");
-                break;
+    protected void doAction(String action) {
+        if (getString(R.string.status_pump).equals(action)) {
+            ListenerService.initiateAction(this, "status pump");
+        } else if (getString(R.string.status_loop).equals(action)) {
+            ListenerService.initiateAction(this, "status loop");
+        } else if (getString(R.string.status_cpp).equals(action)) {
+            ListenerService.initiateAction(this, "opencpp");
+        } else if (getString(R.string.status_tdd).equals(action)) {
+            ListenerService.initiateAction(this, "tddstats");
         }
-
     }
 }
